+++ clauses/make/make_helpers.py
@@ -2,45 +2,18 @@ from ast import literal_eval
 from re import search
 from typing import Union
 
-from exceptions.statements import (
+from exceptions.statements.statements import (
     IllegalNodePropertyType,
     MissingNodeLabel,
     StatementError,
 )
 from models.enums.statement import Statement
 from models.statement import ParsedStatement
-from nodes.node import Node
-from state.wiggle_number import (
-    WIGGLE_NUMBER_FILE_PATH,
-    get_current_wiggle_number,
-    update_wiggle_number,
-)
 from graph_logger.graph_logger import graph_logger
 
 
-def make_node(parsed_statement: ParsedStatement, wiggle_nuber_file_path: str) -> Node:
-
-    if parsed_statement.clause is not Statement.MAKE:
-        raise Exception
-
-    wiggle_number = get_current_wiggle_number(wiggle_nuber_file_path)
-
-    node = Node(
-        wiggle_number=wiggle_number,
-        node_label=parsed_statement.node_label,
-        belongings=parsed_statement.belongings,
-        relations=parsed_statement.relations,
-    )
-    graph_logger.info(f"Created node {wiggle_number}")
-    wiggle_number += 1
-
-    update_wiggle_number(wiggle_nuber_file_path, wiggle_number)
-
-    return node
-
-
 def parse_make_statment(statement_string: str) -> ParsedStatement:
-
+    graph_logger.debug(statement_string)
     # validate that the correct statement is called
 
     statement = None
@@ -160,13 +133,3 @@ def string_to_correct_data_type(value: str) -> Union[float, int, str, list]:
         return value_parsed
     except Exception:
         pass
-
-
-if __name__ == "__main__":
-
-    statement = """MAKE (:NodeLabel{uuid: '7e48f6ae-b25a-4634-91af-b1fb67b90ad9'})"""
-
-    parsed_statment = parse_make_statment(statement)
-
-    x = make_node(parsed_statment, WIGGLE_NUMBER_FILE_PATH)
-    print(x.export_node())
