+++ database/database.json
@@ -1,1 +1,21 @@
-{"3794": {"wiggle_number": 3794, "node_label": "NodeLabel", "created_at": 1666534101.384132, "updated_at": null, "belongings": {"uuid": "7e48f6ae-b25a-4634-91af-b1fb67b90ad9"}, "relations": null}, "9070": {"wiggle_number": 9070, "node_label": "NodeLabel", "created_at": 1666534101.384132, "updated_at": null, "belongings": {"uuid": "7e48f6ae-b25a-4634-91af-b1fb67b90ad9"}, "relations": null}}
+{
+  "50": {
+    "wiggle_number": 50,
+    "node_label": "Node",
+    "created_at": 1666542029.999701,
+    "updated_at": null,
+    "belongings": null,
+    "relations": null
+  },
+  "51": {
+    "wiggle_number": 51,
+    "node_label": "POP",
+    "created_at": 1666542165.340831,
+    "updated_at": null,
+    "belongings": {
+      "first_name": "Harry",
+      "number": 6
+    },
+    "relations": null
+  }
+}
