+++ state/wiggle_number.txt
@@ -1,1 +1,1 @@
-50
+52
